backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
black==24.10.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.3
dill==0.3.9
distlib==0.3.9
distro==1.9.0
fastapi==0.115.2
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
identify==2.6.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jiter==0.6.1
Mako==1.3.5
MarkupSafe==3.0.1
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.1
nodeenv==1.9.1
numpy==2.1.2
openai==1.52.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.3.1
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==75.2.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
sympy==1.13.1
tokenizers==0.20.1
tomlkit==0.13.2
torch==2.5.0
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.0
